requirements.txt
Django==2.2.13
django-mysql==3.6.0
djangorestframework==3.11.0
mysqlclient==1.4.6
python-dateutil==2.8.1
requests==2.23.0
simplejson==3.17.0
python-decouple==3.3
django-extensions==2.2.9
django-jsonfield==1.4.0
pillow==7.1.2
django-rest-swagger==2.2.0

